requirements.txt
asgiref==3.5.2
autopep8==2.0.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.4
Deprecated==1.2.13
Django==4.1.3
django-ckeditor==6.5.1
django-js-asset==2.0.0
django-oauth-toolkit==2.2.0
djangorestframework==3.14.0
idna==3.4
jwcrypto==1.4.2
oauthlib==3.2.2
Pillow==9.3.0
psycopg2==2.9.5
pycodestyle==2.10.0
pycparser==2.21
pytz==2022.6
requests==2.28.1
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.6
urllib3==1.26.13
wrapt==1.14.1
